# lib/api-zod/src/generated/types/adminInquiriesResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminInquiriesResponsePagination } from './adminInquiriesResponsePagination';
import type { AdminInquiry } from './adminInquiry';

export interface AdminInquiriesResponse {
  inquiries: AdminInquiry[];
  pagination: AdminInquiriesResponsePagination;
  unreadCount: number;
}