# lib/api-zod/src/generated/types/adminInquiry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminInquiry {
  id: string;
  fullName: string;
  email: string;
  phone?: string | null;
  affiliation: string;
  inquiryType: string;
  message: string;
  status: string;
  source: string;
  createdAt: string;
  readAt?: string | null;
}