# lib/api-zod/src/generated/types/totpCodeRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TotpCodeRequest {
  /**
     * @minLength 6
     * @maxLength 8
     */
  code: string;
}